requirements.txt
semantic-router[hybrid]>=0.5.0
sentence-transformers>=2.2.0
llama-cpp-python>=0.2.0
huggingface-hub>=0.20.0  # For model downloads
pytest